+++ frontend/src/Components/Pagination/PagedPagination.module.scss
